move `loader.ts` to `platform/`, change many paths
