+++ shells/lib/utils.js
@@ -12,7 +12,7 @@ import {Manifest} from '../../build/runtime/manifest.js';
 import {Arc} from '../../build/runtime/arc.js';
 import {IdGenerator} from '../../build/runtime/id.js';
 import {RecipeResolver} from '../../build/runtime/recipe/recipe-resolver.js';
-import {Loader} from '../../build/runtime/loader.js';
+import {Loader} from '../../build/platform/loader.js';
 import {PecIndustry} from '../../build/platform/pec-industry-web.js';
 import {devtoolsArcInspectorFactory} from '../../build/devtools-connector/devtools-arc-inspector.js';
