+++ src/planning/plan/tests/plan-producer-test.ts
@@ -9,7 +9,7 @@
  */
 import {assert} from '../../../platform/chai-web.js';
 import {Arc} from '../../../runtime/arc.js';
-import {Loader} from '../../../runtime/loader.js';
+import {Loader} from '../../../platform/loader.js';
 import {Manifest} from '../../../runtime/manifest.js';
 import {SingletonStorageProvider} from '../../../runtime/storage/storage-provider-base.js';
 import {FakeSlotComposer} from '../../../runtime/testing/fake-slot-composer.js';
