+++ src/planning/strategies/tests/convert-constraints-to-connections-test.ts
@@ -9,7 +9,7 @@
  */
 import {assert} from '../../../platform/chai-web.js';
 import {Arc} from '../../../runtime/arc.js';
-import {Loader} from '../../../runtime/loader.js';
+import {Loader} from '../../../platform/loader.js';
 import {Manifest} from '../../../runtime/manifest.js';
 import {Modality} from '../../../runtime/modality.js';
 import {FakeSlotComposer} from '../../../runtime/testing/fake-slot-composer.js';
