+++ src/planning/strategies/tests/find-hosted-particle-test.ts
@@ -11,7 +11,7 @@
 
 import {assert} from '../../../platform/chai-web.js';
 import {Arc} from '../../../runtime/arc.js';
-import {Loader} from '../../../runtime/loader.js';
+import {Loader} from '../../../platform/loader.js';
 import {Manifest} from '../../../runtime/manifest.js';
 import {SingletonStorageProvider} from '../../../runtime/storage/storage-provider-base.js';
 import {InterfaceType} from '../../../runtime/type.js';
