+++ src/planning/strategies/tests/map-slots-test.ts
@@ -11,7 +11,7 @@
 
 import {assert} from '../../../platform/chai-web.js';
 import {Arc} from '../../../runtime/arc.js';
-import {Loader} from '../../../runtime/loader.js';
+import {Loader} from '../../../platform/loader.js';
 import {Manifest} from '../../../runtime/manifest.js';
 import {FakeSlotComposer} from '../../../runtime/testing/fake-slot-composer.js';
 import {MapSlots} from '../../strategies/map-slots.js';
@@ -127,7 +127,7 @@ ${recipeManifest}
 
       recipe
         A
-          consume root 
+          consume root
         B
           consume root
         C
