+++ src/planning/tests/simple-planner-test.ts
@@ -11,7 +11,7 @@
 import {assert} from '../../platform/chai-web.js';
 import {checkDefined} from '../../runtime/testing/preconditions.js';
 import {Arc} from '../../runtime/arc.js';
-import {Loader} from '../../runtime/loader.js';
+import {Loader} from '../../platform/loader.js';
 import {Manifest} from '../../runtime/manifest.js';
 import {Match, SimplePlanner} from '../simple-planner.js';
 import {FakeSlotComposer} from '../../runtime/testing/fake-slot-composer.js';
