+++ src/planning/tests/speculator-test.ts
@@ -10,7 +10,7 @@
 
 import {assert} from '../../platform/chai-web.js';
 import {Arc} from '../../runtime/arc.js';
-import {Loader} from '../../runtime/loader.js';
+import {Loader} from '../../platform/loader.js';
 import {Manifest} from '../../runtime/manifest.js';
 import {Speculator} from '../speculator.js';
 import {Id, ArcId} from '../../runtime/id.js';
