+++ src/platform/loader.ts
@@ -1,0 +1,1 @@
+export * from './loader-web.js';
