+++ src/runtime/arc.ts
@@ -13,7 +13,7 @@ import {assert} from '../platform/assert-web.js';
 import {ArcInspector, ArcInspectorFactory} from './arc-inspector.js';
 import {FakePecFactory} from './fake-pec-factory.js';
 import {Id, IdGenerator, ArcId} from './id.js';
-import {Loader} from './loader.js';
+import {Loader} from '../platform/loader.js';
 import {Runnable} from './hot.js';
 import {Manifest} from './manifest.js';
 import {MessagePort} from './message-channel.js';
