--- src/runtime/loader.ts
@@ -1,1 +1,0 @@
-export * from '../platform/loader-web.js';
