+++ src/runtime/manifest.ts
@@ -17,7 +17,7 @@ import {InterfaceInfo} from './interface-info.js';
 import {HandleConnection as InterfaceInfoHandleConnection} from './interface-info.js';
 import {Slot as InterfaceInfoSlot} from './interface-info.js';
 import {Runnable} from './hot.js';
-import {Loader} from './loader.js';
+import {Loader} from '../platform/loader.js';
 import {ManifestMeta} from './manifest-meta.js';
 import * as AstNode from './manifest-ast-nodes.js';
 import {ParticleSpec} from './particle-spec.js';
