+++ src/runtime/manual_tests/loader-test.ts
@@ -11,7 +11,7 @@
 import {assert} from '../../platform/chai-web.js';
 import {fs} from '../../platform/fs-web.js';
 import {path} from '../../platform/path-web.js';
-import {Loader} from '../loader.js';
+import {Loader} from '../../platform/loader.js';
 import {Manifest} from '../manifest.js';
 
 describe('loader', function() {
