+++ src/runtime/tests/multiplexer-test.ts
@@ -11,7 +11,7 @@
 import {assert} from '../../platform/chai-web.js';
 import {Arc} from '../arc.js';
 import {ArcId} from '../id.js';
-import {Loader} from '../loader.js';
+import {Loader} from '../../platform/loader.js';
 import {Manifest} from '../manifest.js';
 import {checkDefined} from '../testing/preconditions.js';
 import {FakeSlotComposer} from '../testing/fake-slot-composer.js';
