+++ src/runtime/tests/particle-interface-loading-with-slots-test.ts
@@ -12,7 +12,7 @@ import {assert} from '../../platform/chai-web.js';
 import {Arc} from '../arc.js';
 import {ArcId} from '../id.js';
 import {HeadlessSlotDomConsumer} from '../headless-slot-dom-consumer.js';
-import {Loader} from '../loader.js';
+import {Loader} from '../../platform/loader.js';
 import {Manifest} from '../manifest.js';
 import {HostedSlotContext, ProvidedSlotContext} from '../slot-context.js';
 import {MockSlotComposer} from '../testing/mock-slot-composer.js';
