+++ src/tests/hotreload-integration-test.ts
@@ -13,7 +13,7 @@ import {StubLoader} from '../runtime/testing/stub-loader.js';
 import {Manifest} from '../runtime/manifest.js';
 import {Arc} from '../runtime/arc.js';
 import {ArcId} from '../runtime/id.js';
-import {Loader} from '../runtime/loader.js';
+import {Loader} from '../platform/loader.js';
 import {FakeSlotComposer} from '../runtime/testing/fake-slot-composer.js';
 import {FakePecFactory} from '../runtime/fake-pec-factory.js';
 import {HeadlessSlotDomConsumer} from '../runtime/headless-slot-dom-consumer.js';
