+++ src/tests/particles/particles-test.ts
@@ -10,7 +10,7 @@
 import {fs} from '../../platform/fs-web.js';
 import {Manifest} from '../../runtime/manifest.js';
 import glob from 'glob';
-import {Loader} from '../../runtime/loader.js';
+import {Loader} from '../../platform/loader.js';
 import {assert} from '../../platform/chai-web.js';
 
 /** Tests that all .schema, .recipe(s) and .manifest files in the particles folder compile successfully. */
