+++ src/tests/slot-composer-test.ts
@@ -12,7 +12,7 @@ import {Planner} from '../planning/arcs-planning.js';
 import {assert} from '../platform/chai-web.js';
 import {Arc} from '../runtime/arc.js';
 import {HeadlessSlotDomConsumer} from '../runtime/headless-slot-dom-consumer.js';
-import {Loader} from '../runtime/loader.js';
+import {Loader} from '../platform/loader.js';
 import {HostedSlotContext, ProvidedSlotContext} from '../runtime/slot-context.js';
 import {MockSlotComposer} from '../runtime/testing/mock-slot-composer.js';
 import {StubLoader} from '../runtime/testing/stub-loader.js';
@@ -231,18 +231,18 @@ recipe
           return class extends DomParticle {
             async setHandles(handles) {
               super.setHandles(handles);
-  
+
               const innerArc = await this.constructInnerArc();
               const hostedSlotId = await innerArc.createSlot(this, 'root');
-        
+
               innerArc.loadRecipe(\`
                 particle A in 'A.js'
                   consume content
                     provide detail
-                
+
                 particle B in 'B.js'
                   consume detail
-                
+
                 recipe
                   slot '\` + hostedSlotId + \`' as hosted
                   A
@@ -252,23 +252,23 @@ recipe
                     consume detail as detail
               \`);
             }
-        
+
             renderHostedSlot(slotName, hostedSlotId, content) {
               this.setState(content);
             }
-        
+
             shouldRender() {
               return Boolean(this.state.template);
             }
-        
+
             getTemplate() {
               return '<div>intercepted-template' + this.state.template + '</div>';
             }
-        
+
             getTemplateName() {
               return this.state.templateName + '/intercepted';
             }
-        
+
             render() {
               return Object.assign({}, this.state.model, {a: this.state.model.a + '/intercepted-model'});
             }
